[bios] Make uses of REAL_CODE() and PHYS_CODE() 64-bit clean
On a 64-bit CPU, any modification of a register by 32-bit or 16-bit
code will destroy the invisible upper 32 bits of the corresponding
64-bit register.  For example: a 32-bit "pushl %eax" followed by a
"popl %eax" will zero the upper half of %rax.  This differs from the
treatment of upper halves of 32-bit registers by 16-bit code: a
"pushw %ax" followed by a "popw %ax" will leave the upper 16 bits of
%eax unmodified.

Inline assembly generated using REAL_CODE() or PHYS_CODE() will
therefore have to preserve the upper halves of all registers, to avoid
clobbering registers that gcc expects to be preserved.

Output operands from REAL_CODE() and PHYS_CODE() assembly may
therefore contain undefined values in the upper 32 bits.

Fix by using explicit variable widths (e.g. uint32_t) for
non-discarded output operands, to ensure that undefined values in the
upper 32 bits of 64-bit registers are ignored.

+++ src/arch/x86/interface/pcbios/memmap.c
@@ -174,7 +174,7 @@ static int meme820 ( struct memory_map *memmap ) {
 	struct memory_region *prev_region = NULL;
 	uint32_t next = 0;
 	uint32_t smap;
-	size_t size;
+	uint32_t size;
 	unsigned int flags;
 	unsigned int discard_D;
 
@@ -216,7 +216,7 @@ static int meme820 ( struct memory_map *memmap ) {
 		}
 
 		if ( size < E820_MIN_SIZE ) {
-			DBG ( "INT 15,e820 returned only %zd bytes\n", size );
+			DBG ( "INT 15,e820 returned only %d bytes\n", size );
 			return -EINVAL;
 		}
